+++ tools/cooja/java/se/sics/cooja/interfaces/ApplicationRadio.java
@@ -26,7 +26,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: ApplicationRadio.java,v 1.6 2008/10/28 12:59:30 fros4943 Exp $
+ * $Id: ApplicationRadio.java,v 1.7 2008/12/04 14:03:42 joxe Exp $
  */
 
 package se.sics.cooja.interfaces;
@@ -59,10 +59,10 @@ public class ApplicationRadio extends Radio implements PolledBeforeActiveTicks {
   private boolean isReceiving = false;
   private boolean isInterfered = false;
 
-  private int transmissionEndTime = 0;
+  private long transmissionEndTime = 0;
 
   private RadioEvent lastEvent = RadioEvent.UNKNOWN;
-  private int lastEventTime = 0;
+  private long lastEventTime = 0;
 
   private boolean outPacketExists = false;
   private RadioPacket outPacket = null;
@@ -126,7 +126,7 @@ public class ApplicationRadio extends Radio implements PolledBeforeActiveTicks {
     return isTransmitting;
   }
 
-  public int getTransmissionEndTime() {
+  public long getTransmissionEndTime() {
     return transmissionEndTime;
   }
 
@@ -217,7 +217,7 @@ public class ApplicationRadio extends Radio implements PolledBeforeActiveTicks {
   }
 
   public void doActionsBeforeTick() {
-    int currentTime = myMote.getSimulation().getSimulationTime();
+    long currentTime = myMote.getSimulation().getSimulationTime();
 
     if (outPacketExists) {
       outPacketExists = false;
